I have added my game that should be fully functional. I made the dice 6 sided instead of 5 sided. I changed variables and methods dealing with money to use/be doubles rather than integers. I also deleted the stupid "Coin Class" file that didn't have any content.

+++ PlayerInterface.java
@@ -1,8 +1,8 @@
 public interface PlayerInterface {
 
-  public void setMoney(int x);
+  public void setMoney(double x);
 
-  public void changeMoney(int x);
+  public void changeMoney(double x);
   
-  public int getMoney();
+  public double getMoney();
 }
